+++ esmerald/routing/gateways.py
@@ -8,6 +8,7 @@ from lilya.routing import Path as LilyaPath, WebSocketPath as LilyaWebSocketPath
 from lilya.types import Receive, Scope, Send
 from typing_extensions import Annotated, Doc
 
+from esmerald.permissions.utils import is_esmerald_permission, wrap_permission
 from esmerald.routing.apis.base import View
 from esmerald.routing.base import Dispatcher
 from esmerald.typing import Void, VoidType
@@ -298,6 +299,14 @@ class Gateway(LilyaPath, Dispatcher, BaseMiddleware, GatewayUtil):
         self._middleware: List["Middleware"] = self.handle_middleware(
             handler=handler, base_middleware=self.middleware
         )
+
+        self.__base_permissions__ = permissions or []
+        self.__lilya_permissions__ = [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
+
         super().__init__(
             path=self.path,
             handler=cast(Callable, handler),
@@ -306,6 +315,7 @@ class Gateway(LilyaPath, Dispatcher, BaseMiddleware, GatewayUtil):
             methods=self.methods,
             middleware=self._middleware,
             exception_handlers=exception_handlers,
+            permissions=self.__lilya_permissions__,  # type: ignore
         )
         """
         A "bridge" to a handler and router mapping functionality.
@@ -317,7 +327,13 @@ class Gateway(LilyaPath, Dispatcher, BaseMiddleware, GatewayUtil):
         self.handler = cast("Callable", handler)
         self.dependencies = dependencies or {}
         self.interceptors: Sequence["Interceptor"] = interceptors or []
-        self.permissions: Sequence["Permission"] = permissions or []  # type: ignore
+
+        self.permissions: Sequence[Permission] = [
+            permission
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]  # type: ignore
+
         self.response_class = None
         self.response_cookies = None
         self.response_headers = None
@@ -523,12 +539,19 @@ class WebSocketGateway(LilyaWebSocketPath, Dispatcher, BaseMiddleware):
         )
         self.is_middleware: bool = False
 
+        self.__base_permissions__ = permissions or []
+        self.__lilya_permissions__ = [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
         super().__init__(
             path=self.path,
             handler=cast("Callable", handler),
             name=name,
             middleware=self._middleware,
             exception_handlers=exception_handlers,
+            permissions=self.__lilya_permissions__,  # type: ignore
         )
         """
         A "bridge" to a handler and router mapping functionality.
@@ -539,7 +562,13 @@ class WebSocketGateway(LilyaWebSocketPath, Dispatcher, BaseMiddleware):
         self.handler = cast("Callable", handler)
         self.dependencies = dependencies or {}
         self.interceptors = interceptors or []
-        self.permissions = permissions or []  # type: ignore
+
+        self.permissions: Sequence[Permission] = [
+            permission
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]  # type: ignore
+
         self.include_in_schema = False
         self.parent = parent
         (handler.path_regex, handler.path_format, handler.param_convertors, _) = compile_path(