+++ esmerald/routing/router.py
@@ -27,6 +27,7 @@ from lilya._internal._module_loading import import_string
 from lilya._internal._path import clean_path
 from lilya.datastructures import URLPath
 from lilya.middleware import DefineMiddleware
+from lilya.permissions import DefinePermission
 from lilya.responses import JSONResponse, Response as LilyaResponse
 from lilya.routing import (
     BasePath as LilyaBasePath,
@@ -56,6 +57,7 @@ from esmerald.interceptors.types import Interceptor
 from esmerald.openapi.datastructures import OpenAPIResponse
 from esmerald.openapi.utils import is_status_code_allowed
 from esmerald.params import Form
+from esmerald.permissions.utils import is_esmerald_permission, wrap_permission
 from esmerald.requests import Request
 from esmerald.responses import Response
 from esmerald.routing._internal import OpenAPIFieldInfoMixin
@@ -481,9 +483,9 @@ class BaseRouter(LilyaRouter):
             path = "/"
         else:
             assert path.startswith("/"), "A path prefix must start with '/'"
-            assert not path.endswith(
-                "/"
-            ), "A path must not end with '/', as the routes will start with '/'"
+            assert not path.endswith("/"), (
+                "A path must not end with '/', as the routes will start with '/'"
+            )
 
         new_routes: list[Any] = []
         for route in routes or []:
@@ -511,9 +513,16 @@ class BaseRouter(LilyaRouter):
                 )
             new_routes.append(route)
 
-        assert lifespan is None or (
-            on_startup is None and on_shutdown is None
-        ), "Use either 'lifespan' or 'on_startup'/'on_shutdown', not both."
+        assert lifespan is None or (on_startup is None and on_shutdown is None), (
+            "Use either 'lifespan' or 'on_startup'/'on_shutdown', not both."
+        )
+
+        self.__base_permissions__ = permissions or []
+        self.__lilya_permissions__ =  [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
 
         super().__init__(
             redirect_slashes=redirect_slashes,
@@ -522,6 +531,7 @@ class BaseRouter(LilyaRouter):
             lifespan=lifespan,
             on_shutdown=on_shutdown,
             on_startup=on_startup,
+            permissions=self.__lilya_permissions__,  # type: ignore
         )
         self.path = path
         self.on_startup = [] if on_startup is None else list(on_startup)
@@ -530,7 +540,13 @@ class BaseRouter(LilyaRouter):
         self.dependencies = dependencies or {}
         self.exception_handlers = exception_handlers or {}
         self.interceptors: Sequence[Interceptor] = interceptors or []
-        self.permissions: Sequence[Permission] = permissions or []  # type: ignore
+
+        self.permissions: Sequence[Permission] = [
+            permission
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]  # type: ignore
+
         self.routes: Any = new_routes
         self.middleware = middleware or []
         self.tags = tags or []
@@ -1988,14 +2004,25 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         """
         if not path:
             path = "/"
+
+        self._application_permissions: Union[Dict[int, Any], VoidType] = Void  # type: ignore
+        self.__base_permissions__ = permissions or []
+
+        self._lilya_permissions: Union[List[DefinePermission], VoidType] = Void
+        self.__lilya_permissions__ = [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
+
         super().__init__(
             path=path,
             handler=handler,
             include_in_schema=include_in_schema,
             exception_handlers=exception_handlers,
             name=name,
+            permissions=self.__lilya_permissions__,  # type: ignore
         )
-
         self._permissions: Union[List[Permission], VoidType] = Void
         self._dependencies: Dependencies = {}
 
@@ -2035,7 +2062,13 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
             description = inspect.cleandoc(self.handler.__doc__ or "") if self.handler else ""
 
         self.description = description
-        self.permissions = list(permissions) if permissions else []  # type: ignore
+
+        self.permissions = [
+            cast(Any, permission)
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]
+
         self.interceptors: Sequence[Interceptor] = []
         self.middleware = list(middleware) if middleware else []
         self.description = self.description.split("\f")[0]
@@ -2061,9 +2094,6 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         if self.responses:
             self.validate_responses(responses=self.responses)
 
-    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> Any:
-        await self.handle_dispatch(scope=scope, receive=receive, send=send)
-
     def validate_responses(self, responses: Dict[int, OpenAPIResponse]) -> None:
         """
         Checks if the responses are valid or raises an exception otherwise.
@@ -2076,6 +2106,9 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
 
             if not is_status_code_allowed(status_code):
                 raise OpenAPIException(detail="The status is not a valid OpenAPI status response.")
+
+    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> Any:
+        await self.handle_dispatch(scope=scope, receive=receive, send=send)
 
     @property
     def http_methods(self) -> List[str]:
@@ -2139,8 +2172,24 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
 
     async def handle_dispatch(self, scope: "Scope", receive: "Receive", send: "Send") -> None:
         """
-        ASGIapp that authorizes the connection and then awaits the handler function.
+        Handles the dispatching of a request.
+        This method processes the incoming request by performing the following steps:
+
+        1. Intercepts the request if interceptors are present.
+        2. Checks if the request method is allowed.
+        3. Creates a Request object from the scope, receive, and send parameters.
+        4. Retrieves the route handler and parameter model for the request method.
+        5. Checks and dispatches application permissions if they exist.
+        6. Gets the response for the request and sends it.
+
+        Args:
+            scope (Scope): The ASGI scope dictionary containing request information.
+            receive (Receive): The ASGI receive callable.
+            send (Send): The ASGI send callable.
+        Returns:
+            None
         """
+
         if self.get_interceptors():
             await self.intercept(scope, receive, send)
 
@@ -2150,9 +2199,16 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         request = Request(scope=scope, receive=receive, send=send)
         route_handler, parameter_model = self.route_map[scope["method"]]
 
-        if self.get_permissions():
+        permissions = self.get_application_permissions()
+        if permissions:
             connection = Connection(scope=scope, receive=receive)
-            await self.allow_connection(connection)
+            # Call dispatcher with the new dispatch_call
+            # To the super
+            dispatch_call = super().handle_dispatch
+
+            await self.dispatch_allow_connection(
+                permissions, connection, scope, receive, send, dispatch_call=dispatch_call
+            )
 
         response = await self.get_response_for_request(
             scope=scope,
@@ -2278,7 +2334,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         return await response_handler(app=app, data=data)  # type: ignore[call-arg]
 
 
-class WebhookHandler(HTTPHandler, OpenAPIFieldInfoMixin, LilyaPath):
+class WebhookHandler(HTTPHandler, OpenAPIFieldInfoMixin):
     """
     Base for a webhook handler.
     """
@@ -2342,6 +2398,7 @@ class WebhookHandler(HTTPHandler, OpenAPIFieldInfoMixin, LilyaPath):
         _path: str = None
         if not path:
             _path = "/"  # pragma: no cover
+
         super().__init__(
             path=_path,
             handler=handler,
@@ -2407,8 +2464,22 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
     ):
         if not path:
             path = "/"
-        super().__init__(path=path, handler=handler, exception_handlers=exception_handlers)
+
+        self._application_permissions: Union[Dict[int, Any], VoidType] = Void  # type: ignore
+        self.__base_permissions__ = permissions or []
+        self.__lilya_permissions__ = [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
+        super().__init__(
+            path=path,
+            handler=handler,
+            exception_handlers=exception_handlers,
+            permissions=self.__lilya_permissions__,  # type: ignore
+        )
         self._permissions: Union[List[Permission], VoidType] = Void
+        self._lilya_permissions: Union[List[DefinePermission], VoidType] = Void
         self._dependencies: Dependencies = {}
         self._response_handler: Union[Callable[[Any], Awaitable[LilyaResponse]], VoidType] = Void
         self._interceptors: Union[List[Interceptor], VoidType] = Void
@@ -2416,7 +2487,13 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
         self.handler = handler
         self.parent: ParentType = None
         self.dependencies = dependencies
-        self.permissions = permissions  # type: ignore
+
+        self.permissions: Sequence[Permission] = [
+            permission
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]  # type: ignore
+
         self.middleware = middleware
         self.signature_model: Optional[Type[SignatureModel]] = None
         self.websocket_parameter_model: Optional[TransformerModel] = None
@@ -2460,13 +2537,33 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
             )
 
     async def handle_dispatch(self, scope: "Scope", receive: "Receive", send: "Send") -> None:
-        """The handle of a websocket"""
+        """
+        Handles the dispatching of a request.
+        This method processes the incoming request by handling interceptors,
+        checking permissions, and dispatching the request to the appropriate
+        handler function.
+
+        Args:
+            scope (Scope): The ASGI scope dictionary containing request information.
+            receive (Receive): The ASGI receive callable to receive messages.
+            send (Send): The ASGI send callable to send messages.
+        Returns:
+            None
+        """
+
         if self.get_interceptors():
             await self.intercept(scope, receive, send)
 
         websocket = WebSocket(scope=scope, receive=receive, send=send)
-        if self.get_permissions():
-            await self.allow_connection(connection=websocket)
+        permissions = self.get_application_permissions()
+        if permissions:
+            # Call dispatcher with the new dispatch_call
+            # To the super
+            dispatch_call = super().handle_dispatch
+
+            await self.dispatch_allow_connection(
+                permissions, websocket, scope, receive, send, dispatch_call=dispatch_call
+            )
 
         kwargs = await self.get_kwargs(websocket=websocket)
 
@@ -2891,6 +2988,12 @@ class Include(LilyaInclude):
         if routes:
             routes = self.resolve_route_path_handler(routes)
 
+        self.__base_permissions__ = permissions or []
+        self.__lilya_permissions__ = [
+            wrap_permission(permission)
+            for permission in self.__base_permissions__ or []
+            if not is_esmerald_permission(permission)
+        ]
         super().__init__(
             path=self.path,
             app=self.app,
@@ -2901,10 +3004,15 @@ class Include(LilyaInclude):
             deprecated=deprecated,
             include_in_schema=include_in_schema,
             redirect_slashes=redirect_slashes,
+            permissions=self.__lilya_permissions__,  # type: ignore
         )
 
         # Making sure Esmerald uses the Esmerald permission system and not Lilya's.
-        self.permissions: Sequence[Permission] = permissions or []  # type: ignore
+        self.permissions: Sequence[Permission] = [
+            permission
+            for permission in self.__base_permissions__ or []
+            if is_esmerald_permission(permission)
+        ]  # type: ignore
 
     def resolve_app_parent(self, app: Optional[Any]) -> Optional[Any]:
         """